requirements.txt
numba
numpy
py-cpuinfo